requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
Pillow==10.1.0
numpy==1.24.3
matplotlib==3.7.2
opencv-python==4.8.1.78
plotly==5.17.0
python-dotenv==1.0.0
